fix: tests & cleanup
