+++ crates/tree-sitter-owl-ms/test/corpus/statements.txt
@@ -204,7 +204,10 @@ Ontology:
 			(sub_class_of
 			  (keyword_sub_class_of)
 				(description
-					(class_iri (abbreviated_iri)))))))
+				  (conjunction
+				    (primary
+					  (class_iri
+                        (abbreviated_iri)))))))))
 
 ===================
 Class Frame SubClassOf with List
@@ -762,60 +765,78 @@ Class: C
 ---
 
 (source_file
-      (prefix_declaration
-        (keyword_prefix)
-        (prefix_name)
-        (full_iri))
-      (ontology
-        (keyword_ontology)
-        (object_property_frame
-          (keyword_object_property)
-          (object_property_iri
-            (simple_iri)))
-        (class_frame
-          (keyword_class)
-          (class_iri
-            (simple_iri)))
-        (class_frame
-          (keyword_class)
-          (class_iri
-            (simple_iri)))
-        (class_frame
-          (keyword_class)
-          (class_iri
-            (simple_iri))
-          (sub_class_of
-            (keyword_sub_class_of)
-            (description
+  (prefix_declaration
+    (keyword_prefix)
+    (prefix_name)
+    (full_iri))
+  (ontology
+    (keyword_ontology)
+    (object_property_frame
+      (keyword_object_property)
+      (object_property_iri
+        (simple_iri)))
+    (class_frame
+      (keyword_class)
+      (class_iri
+        (simple_iri)))
+    (class_frame
+      (keyword_class)
+      (class_iri
+        (simple_iri)))
+    (class_frame
+      (keyword_class)
+      (class_iri
+        (simple_iri))
+      (sub_class_of
+        (keyword_sub_class_of)
+        (description
+          (conjunction
+            (primary
               (object_property_iri
                 (simple_iri))
               (keyword_some)
-              (description
-                (class_iri
-                  (simple_iri))
-                (class_iri
-                  (simple_iri)))))
-          (sub_class_of
-            (keyword_sub_class_of)
-            (description
+              (primary
+                (nested_description
+                  (description
+                    (conjunction
+                      (primary
+                        (class_iri
+                          (simple_iri)))
+                      (primary
+                        (class_iri
+                          (simple_iri)))))))))))
+      (sub_class_of
+        (keyword_sub_class_of)
+        (description
+          (conjunction
+            (primary
               (keyword_inverse)
               (object_property_iri
                 (simple_iri))
               (keyword_some)
-              (description
-                (class_iri
-                  (simple_iri))
-                (class_iri
-                  (simple_iri)))))
-          (comment)
-          (sub_class_of
-            (keyword_sub_class_of)
-            (description
+              (primary
+                (nested_description
+                  (description
+                    (conjunction
+                      (primary
+                        (class_iri
+                          (simple_iri)))
+                      (primary
+                        (class_iri
+                          (simple_iri)))))))))))
+      (comment)
+      (sub_class_of
+        (keyword_sub_class_of)
+        (description
+          (conjunction
+            (primary
               (keyword_inverse)
               (object_property_iri
                 (simple_iri))
               (keyword_some)
+              (primary
+                (class_iri
+                  (simple_iri))))
+            (primary
               (class_iri
-                (simple_iri))
-              (class_iri
-                (simple_iri)))))))
+                (simple_iri)))))))))
